parameterize methods and remove extraneous fields from cross region spec
