+++ test/e2e/resources/cross_region_model.yaml
@@ -9,14 +9,4 @@ spec:
   primaryContainer:
     containerHostname: xgboost
     image: $XGBOOST_V1_IMAGE_URI
-    environment:
-      my_var: my_value
-      my_var2: my_value2
   executionRoleARN: $SAGEMAKER_EXECUTION_ROLE_ARN
-  tags:
-    - key: algorithm
-      value: xgboost
-    - key: environment
-      value: testing
-    - key: customer
-      value: test-user
